Remove the linker script
The final kernel image does not seem to contain any information about
memory layout remaining on it. I understand that it is the Pi bootloader
that places the kernel image on the proper place (0x8000 by default). I
wonder why everyone seems to use a linker script setting the memory
layout then.

A guess, maybe wrong: imagine if I define some "global data" on my code
and refer to it by some label that translates to absolute address. The
tools would need to know where he image would be loaded in order to make
these references point to the right place in memory (as the right place
depends on where the image is loaded).